Stability improvements. Check comment.
Added a check if sensor was broken or disconnected. The “getTemp” function times out after 50ms and returns a huge value (1000) if no pulses are detected. Previous version would get stuck when waiting for a pin state change.

Also added that the timer would stop when not looking for pulses. Should help somewhat with accuracy and stability.